navbar enhancement
