+++ src/components/Body.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import BookList from "./BookList";
-import About from "./About";
-import Home from "./Home";
+import BookList from "./BookList/BookList";
+import About from "./About/About";
+import Home from "./Home/Home";
 
 function Body() {
   return (
