--- src/components/Home.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return (
-    <div>
-      <h1>Welcome to the Book Listing App!</h1>
-      <p>This is the home page.</p>
-    </div>
-  );
-}
