+++ src/components/Home/Home.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <div>
+      Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sapiente et animi iusto porro illum quod, repellendus aperiam dolor ullam saepe? Mollitia repellat repudiandae expedita ratione asperiores dignissimos illo alias atque!
+    </div>
+  );
+}
