+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,29 @@
+.logo{
+    text-transform: uppercase;
+    font-weight: bolder;
+}
+
+.cover{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    width: 90%;
+}
+
+.btn-login{
+    margin-left: auto;
+    margin-top: 0px;
+}
+
+@media (max-width: 991px) {
+    .cover{
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+    }
+    .btn-login{
+        margin-left: 0px;
+        margin-top: 12px;
+    }
+}
