+++ src/components/Navbar/Navbar.js
@@ -1,0 +1,67 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import "./Navbar.css"
+
+export default function Navbar() {
+  return (
+    <nav className="navbar navbar-expand-lg" style={{backgroundColor: '#e3f2fd'}}>
+      <div className="container">
+        <a className="navbar-brand logo" href="/">
+          Informatician
+        </a>
+        <button
+          className="navbar-toggler"
+          type="button"
+          data-bs-toggle="collapse"
+          data-bs-target="#navbarSupportedContent"
+          aria-controls="navbarSupportedContent"
+          aria-expanded="false"
+          aria-label="Toggle navigation"
+        >
+          <span className="navbar-toggler-icon"></span>
+        </button>
+        <div className="collapse navbar-collapse justify-content-center" id="navbarSupportedContent">
+          <div className="cover">
+            <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+              <li className="nav-item">
+                <Link className="nav-link" to="/">
+                  Home
+                </Link>
+              </li>
+              <li className="nav-item">
+                <Link className="nav-link" to="/books">
+                  Books
+                </Link>
+              </li>
+              <li className="nav-item">
+                <Link className="nav-link" to="/about">
+                  About
+                </Link>
+              </li>
+            </ul>
+            <form className="d-flex" role="search" style={{ width: '60%' }}>
+              <input
+                className="form-control me-2"
+                type="search"
+                placeholder="Search"
+                aria-label="Search"
+                style={{ borderRadius: '23px' }}
+              />
+              <button className="btn btn-outline-dark" type="submit">
+                Search
+              </button>
+            </form>
+            <div className="btn-login">
+              <button type="button" class="btn btn-danger">
+                <Link className="nav-link" to="/login">
+                  Sign In
+                </Link>
+              </button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </nav>
+    
+  );
+}
